Refactor: Make secondary indexes optional
