+++ ts/packages/knowPro/src/query.ts
@@ -5,9 +5,7 @@ import {
     DateRange,
     IConversation,
     IMessage,
-    IPropertyToSemanticRefIndex,
     ITermToSemanticRefIndex,
-    ITimestampToTextRangeIndex,
     KnowledgeType,
     ScoredSemanticRef,
     SemanticRef,
@@ -29,10 +27,11 @@ import {
     TermSet,
     TextRangeCollection,
 } from "./collections.js";
-import { PropertyNames } from "./propertyIndex.js";
+import { IPropertyToSemanticRefIndex, PropertyNames } from "./propertyIndex.js";
 import { conversation } from "knowledge-processor";
 import { collections } from "typeagent";
 import { textRangeFromLocation } from "./conversationIndex.js";
+import { ITimestampToTextRangeIndex } from "./timestampIndex.js";
 
 export function isConversationSearchable(conversation: IConversation): boolean {
     return (
